Make PromotionCodeBatchMailer use BaseMailer
All our other mailers inherit from Spree::BaseMailer. ApplicationMailer
is created by default in new rails apps, so it might make sense for
BaseMailer to inherit from that, but I'm not certain it will be defined
in all stores (those upgrading from Rails 4.2). In any case we should be
consistent.