Remove js.erb response for prices#new
This was previously using remote: true and a js.erb template which
replaced the main body of the page with the new form. This just links to
the new url normally. If we wanted the behaviour this had, we should use
turbolinks.

This fixes a minor bug where the NumberWithCurrency JS component wasn't
being initialized on the dynamically added content. It also fixes the
breadcrumb heading.